+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.1.6] - 2026-01-06
+- Added `storage_type` property to `AbstractSSMClient` and `SSMService` and `VaultSSMClient`
+- Improved logging in `MaestroCredentialsService` to include storage type instead of hardcoded "SSM"
+
 ## [7.1.5] - 2025-11-18
 - Enhanced `tracer_decorator` to support functions without `event`/`context` parameters for SRE jobs tracing
 - Update `cachetools` dependency version from `~=6.2.1` to `>=5.5.1,<=6.2.2` for compatibility with Syndicate Rule Engine